// apps/api/src/dto/cart-item.dto.ts
interface ProductDetail {
    categoryImage: {
        mobile: string;
        tablet: string;
        desktop: string;
        external: string;
    };
    new: boolean
    name: string;
    description: string;
    category: string;
    features: string;
    includes: {
        quantity: number;
        item: string;
    }[];
    price: number;
    slug: string;
    quantity: number;
}

export interface CartItemDTO {
    items: [
        {
            product: ProductDetail,
            quantity: number,
        }
    ]
}